Remove MSKTiled and CGFloat arithtmetics
Working on a code restructure.
From now on, MSKTiled can be included separately as an SPM dependency.

+++ MoreSpriteKit/MSKAnimatedLabel.swift
@@ -89,7 +89,8 @@ public class MSKAnimatedLabel: SKNode {
 
     private func addTouchNodeForAutocomplete() {
         let touchNode = SKSpriteNode(color: .clear,
-                                     size: .init(width: labelWidth, height: labels.count * marginVertical))
+                                     size: .init(width: labelWidth, 
+                                                 height: CGFloat(labels.count) * marginVertical))
         touchNode.position.y -= touchNode.frame.height/2
         addChild(touchNode)
     }
